requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
altair==4.2.2
anyio==3.6.2
argilla==1.7.0
async-timeout==4.0.2
attrs==22.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
btrfsutil==6.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
chromadb==0.3.22
click==8.1.3
clickhouse-connect==0.5.24
cmake==3.26.3
colorclass==2.2.2
commonmark==0.9.1
compressed-rtf==1.0.6
crcmod==1.7
cryptography==40.0.2
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
duckdb==0.8.0
easygui==0.98.3
ebcdic==1.1.1
entrypoints==0.4
et-xmlfile==1.1.0
evdev==1.6.0
extract-msg==0.41.1
fastapi==0.95.2
fasttext-wheel==0.9.2
filelock==3.12.0
frozenlist==1.3.1
fsspec==2023.5.0
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
hid==1.0.4
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
idna==3.4
IMAPClient==2.3.1
importlib-metadata==6.6.0
iotop==0.6
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langchain==0.0.166
lark-parser==0.12.0
lit==16.0.5
llama-cpp-python==0.1.48
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
msoffcrypto-tool==5.0.1
multidict==6.0.2
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
oletools==0.60.1
openai==0.27.6
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pandoc==2.3
pcodedmp==1.2.6
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
plumbum==1.8.1
ply==3.11
posthog==3.0.1
progressbar2==4.1.1
protobuf==3.20.3
psutil==5.9.2
pyaml==21.10.1
pyarrow==12.0.0
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.1b0
pyenchant==3.2.2
Pygments==2.15.1
PyGObject==3.42.2
pygpt4all==1.1.0
pygptj==2.0.3
pyinotify==0.9.6
pyllamacpp==2.1.3
Pympler==1.0.1
pypandoc==1.11
pypandoc-binary==1.11
pyparsing==2.4.7
pyrsistent==0.19.3
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
python-utils==3.3.3
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
red-black-tree-mod==1.20
regex==2023.5.5
requests==2.30.0
rfc3986==1.5.0
rich==13.0.1
RTFDE==0.0.2
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
starlette==0.27.0
SteamOS Atomic Updater==0.20190711.0
streamlit==1.22.0
sympy==1.12
systemd-python==235
sysv-ipc==1.1.0
tabulate==0.9.0
team==1.0
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.29.2
triton==2.0.0
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.2
unstructured==0.6.6
urllib3==1.26.6
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
watchdog==3.0.0
watchfiles==0.19.0
websockets==11.0.3
wikipedia==1.4.0
wrapt==1.14.1
XlsxWriter==3.1.0
yarl==1.8.1
zipp==3.15.0
zstandard==0.21.0
